requirements.txt
python-dotenv
pandas
yfinance
sqlalchemy
boto3
psycopg2